tests/test_validation.py
from unittest.mock import mock_open, patch
import unittest
import src.validation as validation


class TestIsValidFilePath(unittest.TestCase):
    def test_unix_path_validation(self):
        test_cases = [
            ("/valid/path", True),
            ("/invalid/path", False),
            # More paths
        ]
        for i, (path, expected) in enumerate(test_cases):
            with self.subTest(i=i):
                result = validation.is_valid_file_path(path)
                self.assertEqual(result, expected)

        # The following is a non-functioning toy example and needs to become a real test.
        m = mock_open(read_data="test data")
        with patch("builtins.open", m):
            result = read_data("any/path/you/want")
        assert result == "test data"
        # End toy example
